Add tweets_to_retrieve option
This adds support for the count parameter in the statuses/user_timeline
api.

In my testing, I found it was necessary to set this to a higher value
than it defaults to in order to get the desired number of tweets to
display when excluding both replies and retweets, and the majority of
recent tweets being one of those types.

From the Twitter API Docs: the acount value is best thought of as a
limit to the number of tweets to return because suspended or deleted
content is removed after the count has been applied. Retweets are also
included in the count, even if include_rts is set false.